Bugfix for: #14 #15 #16 #17
# Fixed issues:
- #14 All contigs that are in the reference but now in coverage data are skipped, and a warning is issued. Furthermore, if no matching contigs are found, Spectre will terminate.
- #15 Due to obsolete code references, Spectre would crash if provided with the `--population` flag in the CNVCaller mode. This has been updated accordingly.
- #16  The flag `--ploidy-chr` required instructions on how to transform the chromosome coverage according to the global ploidy status. This has been changed only to provide the current ploidy state of the chromosome. Spectre will do the transformation calculation automatically now.
- #17 When providing multiple samples without the `--population` flag, results have only been properly outputted to the first sample. Providing multiple samples has only been attended to be used with the `--population` flag. Thus, only CNVs will only be calculated for the first provided sample if the `--population` flag is missing.

# Further updates:
- Fixed: SPC file generation removed all values that were default values according to its datatype.
- Fixed: Re-initialization of variant candidates after loading the SPC file.
- Fixed: Updated functions that used paths only used in the debug mode.
- Added: Two new flags for the population mode `--reciprocal-overlap` and `--disable-quality-filter`:
  - `--reciprocal-overlap` will allow you to change the reciprocal overlap value when looking for the supporting variants in other samples.
  - `--disable-quality-filter` will allow you to look for supporting variants in the variants that did not pass the final filter of Spectre in the CNVCaller mode. (Non-passing reasons of the variants could have been that the variant was too small to pass the minimum CNV length, etc.)
- Added: Pre-calculation checks, which ensure that the user provides proper values.
- Added: Instructions on how to install the latest changes of Spectre locally with pip.
- Updated: Search for supporting variants is primarily done on variants that passed the final filter criteria of the CNVCaller mode.
- Updated: The minimum required package versions.
- Updated: Help message.
- Updated: README.md
- Updated: Minor code changes

# Known issues:
- Population mode: If multiple samples contain the same variant, Spectre will try to find for all of them a supporting variant. This can lead to duplicated variant calls in the final output of the population mode. This will be subject to change in a future release.

+++ spectre/analysis/snv_analysis.py
@@ -128,10 +128,11 @@ class SNVAnalysis(object):
             self.cn_neutral_coverage_stdev = np.nanstd(self.perfect_het_depth)
 
     def het_dump(self):
-        het_dump_file = gzip.open(f'{self.spectre_args.out_dir}/debug/het_dump.txt.gz', 'wt')
-        for het in self.perfect_het_depth:
-            het_dump_file.write(f'{het}\n')
-        het_dump_file.close()
+        if self.spectre_args.as_dev:
+            het_dump_file = gzip.open(f'{self.spectre_args.out_dir}/debug/het_dump.txt.gz', 'wt')
+            for het in self.perfect_het_depth:
+                het_dump_file.write(f'{het}\n')
+            het_dump_file.close()
 
     def het_clean(self):
         # remove top 2% coverage
@@ -506,34 +507,41 @@ class SNVAnalysis(object):
         # -------------------- #
 
     def loh_dump(self, version=""):
-        dump_file = f'{self.spectre_args.out_dir}/debug/loh_dump{version}.tsv'
-        dump_file_bgz = f'{self.spectre_args.out_dir}/debug/loh_dump{version}.tsv.gz'
-        loh_dump_file = open(dump_file, 'w')
-        for lohc in self.loh_candidate_list:
-            loh_dump_file.write(f'{lohc.print(self.cn_neutral_coverage_med, self.ploidy)}\t' \
-                                f'{lohc.filter}\t{lohc.het_score}\t{np.round(lohc.gt_alt_prop, 3)}\n')
-        loh_dump_file.close()
-        pysam.tabix_compress(filename_in=dump_file, filename_out=dump_file_bgz, force=True)
-        pysam.tabix_index(filename=dump_file_bgz, force=True, preset="bed")
-        if os.path.isfile(dump_file_bgz) and os.stat(dump_file_bgz).st_size != 0:
-            os.remove(dump_file)
+        if self.spectre_args.as_dev:
+            dump_file = f'{self.spectre_args.out_dir}/debug/loh_dump{version}.tsv'
+            dump_file_bgz = f'{self.spectre_args.out_dir}/debug/loh_dump{version}.tsv.gz'
+            loh_dump_file = open(dump_file, 'w')
+            for lohc in self.loh_candidate_list:
+                loh_dump_file.write(f'{lohc.print(self.cn_neutral_coverage_med, self.ploidy)}\t' \
+                                    f'{lohc.filter}\t{lohc.het_score}\t{np.round(lohc.gt_alt_prop, 3)}\n')
+            loh_dump_file.close()
+            pysam.tabix_compress(filename_in=dump_file, filename_out=dump_file_bgz, force=True)
+            pysam.tabix_index(filename=dump_file_bgz, force=True, preset="bed")
+            if os.path.isfile(dump_file_bgz) and os.stat(dump_file_bgz).st_size != 0:
+                os.remove(dump_file)
 
     def loh_report(self):
         for lohc in self.loh_candidate_list:
             if lohc.loh_pass:
                 self.logger.debug(lohc.print(self.cn_neutral_coverage_med, self.ploidy))
 
-    def loh_pass_only(self):
+    def loh_pass_and_non_pass(self):
         pass_only = {}
+        fail_only = {}
         for chro in self.spectre_args.only_chr_list:
             if chro not in pass_only:
                 pass_only[chro] = []
+            if chro not in fail_only:
+                fail_only[chro] = []
         for lohc in self.loh_candidate_list:
             if lohc.loh_pass:
                 pass_only[lohc.chromosome].append(lohc)
+            else:
+                fail_only[lohc.chromosome].append(lohc)
         for chro in pass_only:
             self.logger.debug(f'LOH candidates pass {chro}:  {len(pass_only[chro])}')
-        return pass_only
+        return pass_only, fail_only
+
 
     def loh(self, existing_cnv_ids):
         #  > LoH candidates (quick)
